% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.functions.r
\name{is.in}
\alias{is.in}
\title{find matches in values between two vectors}
\usage{
is.in(x, y)
}
\arguments{
\item{x}{vector 1}

\item{y}{vector 2}
}
\description{
find matches in values between two vectors
}
\author{
Marie-Julie Roux
}
