% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.functions.r
\name{mj_boot.a.cat.bycatch}
\alias{mj_boot.a.cat.bycatch}
\title{Calculate catch rate and amount}
\usage{
mj_boot.a.cat.bycatch(cutoff = 50, catchcat = "QMS",
  estimator = "catch.per.tow", do.null = "first.year")
}
\arguments{
\item{cutoff}{not sure}

\item{catchcat}{not sure}

\item{estimator}{not sure}

\item{do.null}{not sure}
}
\description{
FONCTION POUR CALCULER LES VOLUMES ET LES TAUX DE PRISES ACCIDENTELLES EN KG/TRAIT, KG/HR
      OU KG/NO.HOOKS. MJ NOTE: SINCE WE ARE USING THE ADJACENT YEARS (TEMPORAL STRATA) APPROACH TO FILL-IN
      THE GAPS THE "first.stratum" OPTION OF THE "do.null" FUNCTION COMPONENT IS NO LONGER VALID IN THIS
      CODE THE TERMS 'STRATUM' AND 'STRATA' REFER TO SPATIAL SUBDIVISIONS
}
\author{
Marie-Julie Roux
}
