% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.functions.r
\name{mj_boot.effort.group}
\alias{mj_boot.effort.group}
\title{Calculate effort exerted in the by-catch fishery}
\usage{
mj_boot.effort.group(M, obs, effort = "dur2", group = "vesskey",
  G.tol = 0.05)
}
\arguments{
\item{M}{number of bootstrapped estimates required}

\item{obs}{observer data frame (run iteratively for each year and area strata)}

\item{effort}{either a string giving name of effort variable in obs data frame OR if effort = 1 (default) then each row is a unit of effort  if effort a numerical vector of length = dim(obs)[1] then it is the effort}

\item{group}{logical for identifying to which (correlated) group belongs each individual fishing event (unit effort value) eg trip or vessel (If NULL no group quantities tracked, every row its own group. ###TBD)}

\item{G.tol}{level of tolerance in G.ef for accepting boot sample ( default +- 5%)}
}
\value{
a list with components sample, N.ef, G.ef, obs.N.ef, obs.V.ef
     samples list of M vectors of row numbers of obs, one for each bootstrap
     groups list of M vectors of groups sampled, one vector for each bootstrap
     N.ef vector of length M giving the effective sample size (number of effort) for each sample
     G.ef vector of length M giving the effective number of groups (vessels) in each sample
     obs.G.ef scalar, the effective number of groups (vessels) in the obs df
     obs.N.ef scalar the effective number of rows (individual fishing events) in the obs df
}
\description{
FONCTION POUR CALCULER LES VOLUMES ET LES TAUX DE PRISES ACCIDENTELLES EN KG/TRAIT, KG/HR
      OU KG/NO.HOOKS. MJ NOTE: SINCE WE ARE USING THE ADJACENT YEARS (TEMPORAL STRATA) APPROACH TO FILL-IN
      THE GAPS THE "first.stratum" OPTION OF THE "do.null" FUNCTION COMPONENT IS NO LONGER VALID IN THIS
      CODE THE TERMS 'STRATUM' AND 'STRATA' REFER TO SPATIAL SUBDIVISIONS
}
\author{
Marie-Julie Roux
}
